requirements.txt
# Telegram
python-telegram-bot[job-queue]==21.6

# AI
openai==1.54.0

# Memory
mem0ai==0.1.29

# Google Calendar
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Scheduler
APScheduler==3.10.4

# Database
SQLAlchemy==2.0.36

# Web Search
tavily-python==0.5.0
duckduckgo-search==6.3.7

# Utils
python-dotenv==1.0.1
pytz==2024.2
pydantic==2.9.2
httpx==0.27.2
tenacity==9.0.0

# Document Generation (Präsentationen & Tabellen)
python-pptx>=1.0.0
openpyxl>=3.1.0

# TTS – Text-to-Speech (kostenlos, kein API-Key)
gTTS>=2.5.0

# Spotify (optional – nur wenn SPOTIFY_CLIENT_ID gesetzt)
spotipy>=2.24.0

# Logging
structlog==24.4.0

# Dokumenten-Scanner: PDF-Text-Extraktion (optional – Scanner-Service)
PyPDF2>=3.0.0

# Dokument-Scan-Workflow: OCR + durchsuchbare PDFs
pytesseract>=0.3.13
img2pdf>=0.5.1
pypdf>=4.0.0
reportlab>=4.0.0
Pillow>=10.0.0

# Mobilität: Fahrzeiten via OpenRouteService (httpx bereits vorhanden)

# REST API (FastAPI + Uvicorn)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
